[backend/my-item-page]
Modified the names of relationships in Category and User model.

Modified the variable names in homepage and each category page.

Added the link to my item page on photos in homepage and each category page.

Deleted the routes that is unnecessary any more.

Modified ui of my item page and added the links to delete modal on trash icon and edit item page on pen icon.