[backend/each-category-page]
Modified the name of relationships in User and Category model.

Modified the variables in homepage and each category page.

Added the link to my page on photos in homepage and each category page.

Modified ui of my item page.

Added the links to delete modal on trash icon and to edit item page on pen icon.