[backend/each-category-page]
Modified the color of letter form color code to 'black'.

Modified the size of letters and margin top on nav bar.

Modified a word after number of items as it depends on the number.

Modified the position if 'Home'.

Added a space between the number of items and 'item' or 'items'.

Modified the destination of the create category function after category is made in Category Controller.

Added the link of Poppins to import more font weights in app.blade.

Modified the font weight for all in style.css not to change current font weight.

Added delete modal for 'Others' category.

Deleted 'other_item.blade' because it's unnecessary.