[backend/each-category-page]
Added the statement for the function in Register Controller to make categories(Kitchen, Closet, Garage) as default when users are registered.

Modified the Donaiton ID part in my_item.blade.

Modified the ui of homepage to set the Add Category button and Add Item button on the below part of the page.

Modified the avtar on nav bar.